feat: Split up lighting engine tests.
The lighting engine tests have been split up, as the file was too large.
